Fixed view direction reset bug
When clicking in and out of the scene, the view direction would snap
back to an original position. This is handled by a first mouse variable,
which I decided was worth moving out of the camera since the camera
probably shouldn't be handling mouse position code anyway.

Maybe a camera controller is needed?