Reimplemented specular component to phong lighting
I reverted the previous commit implementing this because it was broken.
After a lot of investigation it turns out setting the `u_PointLights` array uniform size to 256 would break the program once the `PointLight` struct in the shader became larger than 9 bytes. I guess there is a limit defined somewhere in OpenGL?